refactor: ensure short-circuit module loads as ES module
