+++ analysis/shortCircuit.mjs
